kogile list delete Finish 190222

+++ backKogile/kogile_view/src/main/java/kogile/checklist/domain/ChecklistVO.java
@@ -5,8 +5,8 @@ import lombok.Data;
 @Data
 public class ChecklistVO {
 	
-	private Long checklist_no;
+	private int checklist_no;
 	private String check_title;
 	private int p_no;
-
+	
 }
